Refactor product list to it's own component
